Use new store function in repository crate for player_number setting
