+++ crates/repository/Cargo.toml
@@ -11,6 +11,7 @@ constants = { workspace = true }
 glob = { workspace = true }
 futures-util = { workspace = true }
 home = { workspace = true }
+parameters = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 spl_network_messages = { workspace = true }
